Add user creation backend feature
Users can be created if a post request is sent to
"localhost:3001/api/users" with JSON body attached to the request
object. The JSON body should contain an email address and a password.
The provided email should not be present in the DB before creating a new
user.

Currently the sender provided email and password are blindly added to
the DB without checking if they satisfy any constraints.