Create basic checkout page
The shopping cart items are queried individually using the fake store
api due to which the checkout page takes time to load. I might want to
save the data provided by the fake api inside my DB so that I can
query faster using populate method of mongoose. It is not an issue
I want to worry about right now so I'll leave it as it is for now.