Add shopping cart manipulation to the database
The quantity of each product in the shopping cart is now stored in the
database, making the addition of products and changing the quantity
very controlled. The application has slowed down because of the database
values changing everytime during checkout but its fine. The "info.txt"
file is not consistent with the changes made yet.