Add "name" property in the users documents
The "name" field was added to the documents of "users" collection.
The value associated to this field will be utilized when a user places
an order and for UI. The "name" field should be at least 3 characters
long is the constraint I added to act as a good starting point for
further validations of this field.